TM_server/requirements.txt
Django==2.1.2
django-cleanup==0.1.9
djangorestframework==3.9.3
django-cors-headers==2.5.3
CairoSVG==2.4.0
cairocffi==1.0.2
cffi==1.12.3